Experiments/IdentityDetection/PreprocessingScripts/requirements.txt
alabaster==0.7.12
appdirs==1.4.3
asn1crypto==0.24.0
autopep8==1.3.4
Babel==2.8.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.6.0
cachetools==3.1.1
certifi==2018.4.16
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
cryptography==2.1.4
cycler==0.10.0
Cython==0.29.14
decorator==4.3.0
docutils==0.16
dominate==2.3.1
drive-cli==2.1.0
easydict==1.9
enum34==1.1.6
face-alignment==1.0.0
filelock==3.0.12
freetype-py==2.1.0.post1
future==0.18.2
gdown==3.12.2
google-api-python-client==1.7.8
google-auth==1.22.1
google-auth-httplib2==0.0.4
graphviz==0.8.4
html5lib==0.999999999
httplib2==0.12.1
idna==2.6
imageio==2.3.0
imagesize==1.2.0
insightface==0.1.5
ipaddress==1.0.17
Jinja2==2.11.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
lxml==4.2.1
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==2.2.4
moviepy==0.2.3.4
mxnet-cu102==1.6.0.post0
mxnet-cu80==1.5.0
networkx==2.2
numpy==1.16.6
oauth2client==4.1.3
opencv-python==3.4.0.12
packaging==20.1
pep8==1.7.1
pick==0.6.4
Pillow==5.1.0
prettytable==0.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.4.3
pycairo==1.16.2
pycodestyle==2.3.1
pycrypto==2.6.1
pyfiglet==0.8.post1
pyglet==1.4.0b1
Pygments==2.5.2
PyOpenGL==3.1.0
pyparsing==2.4.6
PyPDF2==1.26.0
pyrender==0.1.33
PySocks==1.7.1
python-dateutil==2.8.1
pytools==2020.1
pytz==2017.3
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==5.3
pyzmq==17.0.0
requests==2.21.0
rsa==4.5
scikit-image==0.14.5
scikit-learn==0.20.4
scipy==1.2.2
scour==0.36
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==2.0.0
Sphinx==1.8.5
sphinx-bootstrap-theme==0.7.1
sphinxcontrib-websupport==1.1.2
subprocess32==3.5.4
torch==0.4.0
torchfile==0.1.0
tornado==5.0.2
tqdm==4.23.3
trimesh==3.5.16
typing==3.7.4.1
uritemplate==3.0.1
urllib3==1.22
visdom==0.1.7
webencodings==0.5
